move some entities
