+++ src/dishka/__init__.py
@@ -8,14 +8,13 @@ __all__ = [
 ]
 
 from .async_container import AsyncContainer, make_async_container
-from .component import DEFAULT_COMPONENT, Component
 from .container import Container, make_container
 from .dependency_source import (
-    DependencyKey,
-    FromComponent,
     alias,
     decorate,
     provide,
 )
+from .entities.component import DEFAULT_COMPONENT, Component
+from .entities.key import DependencyKey, FromComponent
+from .entities.scope import BaseScope, Scope
 from .provider import Provider
-from .scope import BaseScope, Scope
