+++ src/dishka/dependency_source/__init__.py
@@ -3,12 +3,10 @@ __all__ = [
     "alias", "Alias",
     "decorate", "Decorator",
     "provide", "Factory", "FactoryType",
-    "DependencyKey", "FromComponent",
 ]
 
 from .alias import Alias, alias
 from .decorator import Decorator, decorate
 from .factory import Factory, FactoryType, provide
-from .key import DependencyKey, FromComponent
 
 DependencySource = Alias | Factory | Decorator
