+++ src/dishka/entities/component.py
