+++ src/dishka/entities/scope.py
