+++ tests/unit/container/test_components.py
@@ -2,8 +2,7 @@ from typing import Annotated
 
 import pytest
 
-from dishka import Provider, Scope, make_container, provide
-from dishka.dependency_source.key import FromComponent
+from dishka import FromComponent, Provider, Scope, make_container, provide
 from dishka.exceptions import NoFactoryError
